Moved Properties from Entitas.Unity to Entitas.Serialization
