--- Entitas.Unity/Assets/Entitas/Unity/Properties.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0f116ee381a8645bc9265e203b09c0fa
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
